requirements.txt
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==2.1.5
matplotlib==3.7.2
mistune==3.0.2
numpy==1.24.3
packaging==24.1
pandas==2.0.3
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
scikit-learn==1.3.0
six==1.16.0
SQLAlchemy==2.0.34
tensorflow==2.13.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
joblib==1.3.2